Split plugin modules and add tests
